requirements.txt
# Streamlit for the web UI
streamlit==1.48.0


# Requests for making HTTP API calls
requests>=2.31.0


# python-docx to create Word documents
python-docx>=0.8.12


# FPDF to create PDF files
fpdf>=1.7.2


# Optional: OpenAI client for OpenRouter calls
openai>=0.27.9


# python-dotenv>=1.0.0